requirements.txt
# Core
python-dotenv>=1.0.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# ML
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Dashboard
flask>=2.3.0

# Testing
pytest>=7.4.0
